+++ tests/Unit/Api/News/ListByProjectTest.php
@@ -1,0 +1,78 @@
+<?php
+
+namespace Redmine\Tests\Unit\Api\News;
+
+use PHPUnit\Framework\TestCase;
+use Redmine\Api\News;
+use Redmine\Client\Client;
+
+/**
+ * Tests for News::listByProject()
+ */
+class ListByProjectTest extends TestCase
+{
+    /**
+     * @covers \Redmine\Api\News::listByProject
+     */
+    public function testListByProjectWithoutParametersReturnsResponse()
+    {
+        // Test values
+        $projectId = 5;
+        $response = '["API Response"]';
+        $expectedReturn = ['API Response'];
+
+        // Create the used mock objects
+        $client = $this->createMock(Client::class);
+        $client->expects($this->once())
+            ->method('requestGet')
+            ->with(
+                $this->stringStartsWith('/projects/5/news.json')
+            )
+            ->willReturn(true);
+        $client->expects($this->exactly(1))
+            ->method('getLastResponseBody')
+            ->willReturn($response);
+        $client->expects($this->exactly(1))
+            ->method('getLastResponseContentType')
+            ->willReturn('application/json');
+
+        // Create the object under test
+        $api = new News($client);
+
+        // Perform the tests
+        $this->assertSame($expectedReturn, $api->listByProject($projectId));
+    }
+
+    /**
+     * @covers \Redmine\Api\News::listByProject
+     */
+    public function testListByProjectWithParametersReturnsResponse()
+    {
+        // Test values
+        $projectId = 5;
+        $parameters = ['not-used'];
+        $response = '["API Response"]';
+        $expectedReturn = ['API Response'];
+
+        // Create the used mock objects
+        $client = $this->createMock(Client::class);
+        $client->expects($this->once())
+            ->method('requestGet')
+            ->with(
+                $this->stringContains('not-used')
+            )
+            ->willReturn(true);
+        $client->expects($this->exactly(1))
+            ->method('getLastResponseBody')
+            ->willReturn($response);
+        $client->expects($this->exactly(1))
+            ->method('getLastResponseContentType')
+            ->willReturn('application/json');
+
+        // Create the object under test
+        $api = new News($client);
+
+        // Perform the tests
+        $this->assertSame($expectedReturn, $api->listByProject($projectId, $parameters));
+    }
+}